+++ sha512.hpp
@@ -76,8 +76,8 @@ namespace hmac_hash {
 
     protected:
         void transform(const uint8_t *message, size_t block_nb);
-        size_t m_tot_len;
-        size_t m_len;
+        uint64_t m_tot_len;
+        uint64_t m_len;
         uint8_t m_block[2 * SHA384_512_BLOCK_SIZE];
         uint64_t m_h[8];
     };
